// CTT_Online_Shopping/FrontEnd/public-portal/src/app/models/enums.ts
export enum FilterAttributeType {
  String = 'String',
  Boolean = 'Boolean',
  Number = 'Number',
}
export enum VariantType {
  Color = 'Color',
  Size = 'Size',
  ColorAndSize = 'ColorAndSize',
}
// enums
export enum DiscountType {
  None = 'None',
  Fixed = 'Fixed',
  Percentage = 'Percentage',
}